add template class
