+++ lib/sheldon/template.rb
@@ -1,0 +1,6 @@
+class Sheldon
+
+  class Template
+  end
+
+end
