+++ spec/sheldon/template_spec.rb
@@ -1,0 +1,8 @@
+require File.expand_path('../../spec_helper', __FILE__)
+
+describe Sheldon::Template do
+
+  it 'is ok' do
+  end
+
+end
